Fix misalignment issues by implementing hybrid Arena/BumpAllocator
This fix required deep debugging, including reviewing memory layout in the TRM.
The new allocator is a hybrid mix of an Arena and a BumpAllocator.
While not the most memory-efficient, it is extremely fast and only frees memory
when everything has been dropped.

With up to 4MiB of SRAM, delayed memory freeing is negligible.
**Important for future use:** Always set the  feature flag,
or external RAM will not be mapped into the address space.
Safety commit